src/mlproject01/utils/common.py
import os
from box.exceptions import BoxValueError
import yaml
from mlproject01 import logger
import json
import joblib 
from ensure import ensure_annotations
from box import ConfigBox
from pathlib import Path
from typing import Any


@ensure_annotations
def read_yaml(yaml_path: Path) -> ConfigBox:
    pass

@ensure_annotations
def create_directories(directory_paths: list, verbose=True):
    pass

@ensure_annotations
def load_json_file(path: Path) -> ConfigBox:
    pass

@ensure_annotations
def save_json(path: Path, data: dict):
    pass

@ensure_annotations
def load_bin_file(path: Path) -> Any:
    pass

@ensure_annotations
def save_bin(data: Any, path: Path):
    pass

@ensure_annotations
def get_size(path: Path) -> str:
    pass
